* Remove unused function

* remove uref cep47 event version

# Conflicts:
#	contract/src/events.rs
#	contract/src/main.rs

* Approve test

* Add one test for meta data update in CEP78

* Bad errors

* Lint

* debug

* remove comments

* Approve all to Approve

+++ tests/src/installer.rs
@@ -331,7 +331,7 @@ fn should_prevent_installation_with_ownership_and_minting_modality_conflict() {
         .with_collection_name(NFT_TEST_COLLECTION.to_string())
         .with_collection_symbol(NFT_TEST_SYMBOL.to_string())
         .with_total_token_supply(1_000u64)
-        .with_minting_mode(MintingMode::Installer as u8)
+        .with_minting_mode(MintingMode::Installer)
         .with_ownership_mode(OwnershipMode::Minter)
         .with_reporting_mode(OwnerReverseLookupMode::Complete)
         .build();
